chg: RunCommand In CLI Package
RunCommand parameters have changed to take the working directory as the
first argument. Arguments for the command have been change to an array
of strings. 2 new parameters for env setting the sub process environment
variables. And the 2 previous functions have merged into 1 so that it
now takes input.

BREAKING CHANGE