rmv: All Lowercase Variables From cli Usage
Updated the methods to remove all of the lowercase variable placeholders
from the default UsageTmpl in the cli package.
This was partly done in version 2.5.2. which was incorrectly marked as a
patch when it should have been a major release.

That version will be removed since its been out less that an hour.

BREAKING CHANGE