fix: Session Manager Remove Feature
The Remove method of the session manager has been changed to
return an error when it cannot find the key in the session.
Before it blindly tried to delete the session data even if there
was no key in the session.

BREAKING CHANGE: The Remove method of the session manager now returns an error.