chg: Session ID Getter
Change the ID method to return a string instead of a HTTP
cookie. Added a new function session.Manager.IDCookie to perform
the previous functionality. This allows the session ID to be
used in other contexts, such as in a URL or as a header value.
This change also makes it clear to the developer.

BREAKING CHANGE: Session ID is now a string instead of a HTTP
cookie. The session.Manager.IDCookie function can be used to
return the previous functionality.