chg: web.Response To support HTTP Headers
Changed the headers type to be compatible with the http.response
header type. Adding a necessary conversion method for Lambda to
maintain existing functionality.

BREAKING CHANGE: The web.Response.Headers type has been changed
from a map[string]string to a map[string][]string. This change
is necessary to support the HTTP protocol and is not compatible
with the previous version of the web.Response type. If you are
using the web.Response type in your code, you will need to
update your code to use the new type.