chg: Web Response To Implement Response Writer
Changed the web.Response type to also implemnt the http.ResponseWriter
interface. This allows, for example, the use of http.SetCookie
function to write cookies to the response.
This change is part of a larger effort to make the web package
more compatible with the standard library's net/http package. While
also maintaining the existing functionality with AWS Lambda packages.