add: Request To Web Package
Added a complimentary `Request` struct to the `web` package. This struct
is designed to encapsulate the HTTP request data, and acting as a
medium between other types of request. This struct is intended to be
used in conjunction with the `web` package's `Response` struct to
facilitate the handling of HTTP requests and responses in a more
structured and organized manner.