chg: Finish Wrapping http.Request
Update the web.Request to wrap http.Request. This included ensuring
that events.LambdaFunctionUrlRequest is properly translated into
an http.Request. At least the POST body for now.

This is a wip commit. I need to add more tests and ensure that
the web.Request is properly wrapped in all cases.