rmv: Web and Session To It Own Library
While initially it was a good fit, as the project grew, it became
clear that it stood on its own. This commit moves the web library
to its own repository at github.com/kohirens/www.

BREAKING CHANGE: The web library has been moved to its own
repository. If you are using the web library, you will need to
update your import paths to point to the new repository.